Moved background.html and friends into their own folder
