+++ addon/background/uiHandler.js
@@ -2,7 +2,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at https://mozilla.org/MPL/2.0/. */
 
-import { browser as CompatBrowser } from "./content/es-modules/thunderbird-compat.js";
+import { browser as CompatBrowser } from "../content/es-modules/thunderbird-compat.js";
 
 if (!globalThis.browser) {
   globalThis.browser = CompatBrowser;
