+++ addon/background/window.js
