+++ addon/content/es-modules/utils.js
@@ -26,32 +26,6 @@ export function getInitials(name) {
   return initials.toUpperCase();
 }
 
-/**
- * Hash an email address to produce a color. The same email address will
- * always return the same color.
- *
- * @param {string} email
- * @returns {string} - valid css hsl(...) string
- */
-export function freshColor(email) {
-  let hash = 0;
-  for (let i = 0; i < email.length; i++) {
-    let chr = email.charCodeAt(i);
-    hash = (hash << 5) - hash + chr;
-    hash &= 0xffff;
-  }
-  let hue = Math.floor((360 * hash) / 0xffff);
-
-  // try to provide a consistent lightness across hues
-  let lightnessStops = [48, 25, 28, 27, 62, 42];
-  let j = Math.floor(hue / 60);
-  let l1 = lightnessStops[j];
-  let l2 = lightnessStops[(j + 1) % 6];
-  let lightness = Math.floor((hue / 60 - j) * (l2 - l1) + l1);
-
-  return "hsl(" + hue + ", 70%, " + Math.floor(lightness) + "%)";
-}
-
 // Taken from
 // https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/String/charAt#Fixing_charAt()_to_support_non-Basic-Multilingual-Plane_(BMP)_characters
 function fixedCharAt(str, idx) {
