+++ addon/tests/uiHandler.test.jsx
@@ -7,7 +7,7 @@ import { jest } from "@jest/globals";
 // This is so that utils.js defines fetch onto the global scope.
 // eslint-disable-next-line no-unused-vars
 import { enzyme } from "./utils.js";
-import { UIHandler } from "../uiHandler.js";
+import { UIHandler } from "../background/uiHandler.js";
 
 describe("getDefaultIdentity", () => {
   let uiHandler;
