+++ webpack.config.cjs
@@ -10,13 +10,14 @@ module.exports = {
     gallery: "./addon/gallery/gallery.jsx",
     options: "./addon/options/options-render.js",
     stub: "./addon/content/stub.js",
+    background: "./addon/background/background.js",
     "dev-frame": "./addon/dev-frame/dev-frame-render.js",
   },
   mode: "none",
   optimization: {
     minimize: false,
     splitChunks: {
-      name: "common",
+      name: false,
       chunks: "all",
       minChunks: 1,
     },
@@ -49,6 +50,12 @@ module.exports = {
       template: "./addon/content/stub.html",
       chunks: ["stub"],
       filename: "stub.html",
+    }),
+    new HtmlWebpackPlugin({
+      hash: false,
+      template: "./addon/background/background.html",
+      chunks: ["background"],
+      filename: "../background/background.html",
     }),
     new HtmlWebpackPlugin({
       hash: false,
